requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# LangChain
langchain==0.0.354
langchain-community==0.0.16
langchain-core==0.1.16
langchain-google-genai==0.0.6

# Document Processing
pypdf==3.17.1
docx2txt==0.8
python-docx==1.1.0
python-magic==0.4.27
unstructured==0.10.30
unstructured-inference==0.7.21
unstructured-docx==0.0.3
unstructured-pdf==0.1.2
unstructured-doctools==0.0.4

# Additional Dependencies
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.0
reportlab==4.0.9
pytesseract==0.3.10
python-pptx==0.6.23
